Move home api response type to home api file
