+++ apis/homeApi.ts
@@ -1,6 +1,30 @@
 // import fetchData from "@/utils/util-func";
-import { HomeResponse, RecruitArticlesResponse } from "@/types/home";
 // import studium from "./paths";
+
+// 최초 홈 진입 통 API
+// https://api.server.d0lim.com/studium/v1/home
+
+export type HomeResponse = {
+  popularRecruitArticles: HomeRecruitArticleResponse[];
+} & RecruitArticlesResponse;
+
+// 추가 게시글 요청
+// https://api.server.d0lim.com/studium/v1/home/article?last={oldest_recent_article_id}&size={size}&sort={sortBy}
+export type RecruitArticlesResponse = {
+  recruitArticles: HomeRecruitArticleResponse[];
+  lastRecruitArticleId: number;
+};
+
+type HomeRecruitArticleResponse = {
+  id: number;
+  title: string;
+  description: string;
+  tags: Tag[];
+  createdAt: Date;
+  expiresAt: Date;
+};
+
+// https://api.server.d0lim.com/studium/v1/recruit/{id}
 
 const createRecruitItem = (count: number) =>
   Array.from({ length: count }, (v, i) => i + 1).map(v => ({
